Major Changes to the Interface and Code Validation
Now using mso folder browser to select folder for origin.
Now using a datagridview to show all PDF files in origin directory.
Changed schema
Removed Heat textbox
Updated the implementation of the version info in the title bar to get
rid of the unhandled exception on open.
Added validationto check for PDF's in selected origin
Added validation on WriteHeat_Click to ensure a directory has been
provided for the origin
Added validation on AriteHeat_Click to ensure a directory has been
provided for the destination.
Updated WriteHeat_Click to make the newFile name include the heat number
instead of "with heat".
Updated form load to uneneable the "WriteHeat button - this will only be
enabled when both directories have valid values.